Change non-archival references to use citeNA variants. Change references to attached documents to use Appendix references
Non-archival references now use citeNA variants.
References to Naroff and Watt 1987 switched from a citation to an
in-document section reference to Appendix B.
References to attached spec.language and spec.runtime documents now
instead are section references to Appendix D and E.
Page numbers referenced in these Appendices now use the document global
page numbers instead of their original page numbers.
These numbers are entered as relative offsets from the Appendix
document’s start page number.
This uses the pageref package to convert the pageref (which is a
string) back to a number so numbers can be added to it.