White out top part of spec.language/spec.runtime
Also changed Objective-C vs C++ subsection back to \subsection from
\paragraph but with * to omit the number.
Also modified page range of Appendix B citation to account for the fact
that only odd numbered pages are included, so the range ends on page 39
instead of 41.

Notes on the solution I came up with for the problem that the tops of
the spec.language and spec.runtime PDFs were dirty and overlapping with
the Appendix section header. I couldn’t get picturecommand to work
properly; attempts to move the origin of the rectangle drawn with the
tikz package just ended up making the rectangle bigger for some reason.
Moving the rectangle fill command to pagecommand gave me almost what I
wanted, but unfortunately using pagecommand forced it to draw the white
rectangle after the section header as it now assumed it was part of the
text. This still left a sliver of the dirty scan underneath. Lacking
any other solution, I had to shorten the Appendix section titles so
that they would fit in one line. This kind of makes drawing the white
rectangle not as necessary, since the overlap no longer occurs anyway,
and there’s no way to get the white rectangle to draw under the section
title when using pagecommand. But at least it does clean up the dirtier
section of the scan. The only loss is the loss of the full title of the
two documents in the appendix section titles, which is also displayed
in the table of contents. I chose to use the short “spec.language” and
“spec.runtime” titles in the actual section titles in order to maintain
the S.Naroff c.1987 attribution and the “Historical document” prefix to
be consistent with the other appendices.

I also had attempted to modify the Pdfs themselves by drawing a white
box in them in Preview. But for some reason either Overleaf has
permanently cached the old versions, or it somehow discards the added
white box, so that doesn’t work. I don’t have Acrobat Pro so I can’t
flatten the file to see if that fixes it, nor do I have Photoshop to
allow me to simply erase the dirty sections of the PDF. But I suspect
that Overleaf is caching the old file as even when I deleted the files
from Overleaf it would still compile with the PDFs as before!
In any case, this is the reason I changed the filenames from
SNaroffSpec_runtime.pdf to SNaroffSpec-runtime.pdf, to try to force
Overleaf to use the new versions of the files. That didn’t work, making
me think that it’s simply discarding the white box as an annotation and
not an actual part of the PDF.